# FORUM_OPENAPI_SPEC.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DAIP-LIVE MVP API",
    "description": "API for multi-role debate, agile project execution, and hallucination suppression. V0.3.12 with three core scenarios: Expert Consultation, Academic Research, and Industry Analysis, plus Forum Mode.",
    "version": "0.3.12"
  },
  "paths": {
...
    "/api/forum/session/{session_id}/messages": {
      "get": {
        "tags": ["forum"],
        "summary": "Get Session Messages",
        "description": "Get message history for a specific session.",
        "operationId": "get_session_messages",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "session_id": {"type": "string"},
                    "messages": {"type": "array", "items": {"type": "object"}},
                    "message_count": {"type": "integer"},
                    "timestamp": {"type": "string", "format": "date-time"}
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        },
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "loc": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "msg": {"type": "string"},
          "type": {"type": "string"}
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}